added and configured DOTENV
