+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "weather-app",
   "version": "1.0.0",
-  "description": "Weather Application",
-  "main": "src/index.js",
+  "description": "[Solo Project]\r Javascript",
+  "main": "index.js",
   "scripts": {
-    "build": "webpack --mode development",
     "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack --mode development",
     "watch": "webpack --progress --watch"
   },
   "repository": {
@@ -19,15 +19,14 @@
   },
   "homepage": "https://github.com/geraldgsh/weather-app#readme",
   "dependencies": {
-    "lodash": "^4.17.15",
-    "webpack": "^4.41.6",
-    "webpack-cli": "^3.3.11",
-    "watch": "webpack --progress --watch"
+    "dotenv": "^8.2.0",
+    "dotenv-webpack": "^1.7.0",
+    "webpack": "^4.42.0",
+    "webpack-cli": "^3.3.11"
   },
   "devDependencies": {
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-plugin-import": "^2.20.1",
-    "url-loader": "^3.0.0"
+    "eslint-plugin-import": "^2.20.1"
   }
 }
