+++ src/modules/control/localControl.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-undef */
 /* eslint-disable no-console */
 import { renderLocalCard, localCelsius, localFahrenheit } from '../DOM/localDOM';
 import { cityEntry, CurrentCity } from '../constructor/localConstructor';
@@ -9,7 +10,7 @@ const updateLocalStorage = (arr) => {
 const fetchLocalCityWeather = (cityName, unit) => {
   const location = cityName;
   const baseUrl = 'https://api.openweathermap.org/data/2.5';
-  const key = '05f63ad5080a502f607cfa5b1219794b';
+  const key = process.env.OPENKEY;
   const value = unit;
   const url = `${baseUrl}/weather?q=${location}&units=${value}&APPID=${key}`;
   fetch(url, { mode: 'cors' })
@@ -44,7 +45,8 @@ const addCurrentCity = (city, unit) => {
 };
 
 const fetchLocalCityName = (unit) => {
-  fetch('https://ipinfo.io?token=311875147c07c2', { mode: 'cors' })
+  const KEY = process.env.IPKEY;
+  fetch(`https://ipinfo.io?token=${KEY}`, { mode: 'cors' })
     .then((response) => response.json())
     .then((response) => {
       addCurrentCity(response.city, unit);
