+++ webpack.config.js
@@ -1,6 +1,10 @@
 const path = require('path');
+const Dotenv = require('dotenv-webpack');
 
 module.exports = {
+  plugins: [
+    new Dotenv(),
+  ],
   mode: 'development',
   entry: './src/index.js',
   output: {
